requirements.txt
# Core deep learning frameworks
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Computer vision and image processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Data handling and numerical computing
numpy>=1.24.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bars and utilities
tqdm>=4.65.0

# Configuration and serialization
PyYAML>=6.0

# Scientific computing
scipy>=1.10.0
scikit-learn>=1.3.0

# Optional: For better performance
# Install CUDA-enabled PyTorch from https://pytorch.org/get-started/locally/

